backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyyaml
groq
langgraph
pinecone
tavily-python
httpx
aiosqlite
prometheus-fastapi-instrumentator
prometheus-client
pytest
pytest-asyncio
